# feat(query): support early filtering for hash join