I hereby agree to the terms of the CLA available at: https://databend.rs/dev/policies/cla/

## Summary

Benchmark TPC-H SF100:
```
Q3:  2.1s -> 1.9s
Q4:  2.3s -> 2.1s
Q5:  4.2s -> 3.9s
Q12: 3.3s -> 3.1s
Q21: 7.7s -> 7.0s
Q10: 4.0s -> 3.8s
```

- Closes #issue